pipeline fix

+++ components/automate-cli/cmd/chef-automate/migration_pg.go
@@ -900,7 +900,7 @@ func calDiskSizeAndDirSize(data, oldData string) (bool, error) {
 	dirSizeInMb := size / (1024 * 1024)
 
 	msg := fmt.Sprintf("Insufficient disk space for migration.\n%s: %5d MB\n%s: %5d MB\n%s",
-		"Space Required", (dirSizeInMb + int64(dirSizeInMb/10)),
+		"Space Required", (dirSizeInMb + (dirSizeInMb / 10)),
 		"Space Available", diskSpaceInMb,
 		"To continue with less memory Please use --skip-storage-check")
 	// NewData > olddata + 10%of oldData
